# restapi/server.go
// Package Swagger provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package restapi

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	externalRef0 "github.com/gofiber/fiber/v2"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /articles)
	GetArticles(ctx *externalRef0.Ctx) error

	// (POST /articles)
	CreateArticle(ctx *externalRef0.Ctx) error
	// Deletes an article
	// (DELETE /articles/{slug})
	Delete(ctx *externalRef0.Ctx, slug string) error

	// (GET /articles/{slug})
	GetArticleBySlug(ctx *externalRef0.Ctx, slug string) error
	// Updates an article
	// (PUT /articles/{slug})
	UpdateArticle(ctx *externalRef0.Ctx, slug string) error
	// Create an author
	// (POST /authors)
	CreateAuthor(ctx *externalRef0.Ctx) error
	// Find article by slug
	// (GET /authors/{name})
	GetAuthorByName(ctx *externalRef0.Ctx, name string) error

	// (GET /tags)
	GetTags(ctx *externalRef0.Ctx) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetArticles converts echo context to params.
func (w *ServerInterfaceWrapper) GetArticles(ctx *externalRef0.Ctx) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetArticles(ctx)
	return err
}

// CreateArticle converts echo context to params.
func (w *ServerInterfaceWrapper) CreateArticle(ctx *externalRef0.Ctx) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.CreateArticle(ctx)
	return err
}

// Delete converts echo context to params.
func (w *ServerInterfaceWrapper) Delete(ctx *externalRef0.Ctx) error {
	var err error
	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameter("simple", false, "slug", ctx.Params("slug"), &slug)
	if err != nil {
		return externalRef0.NewError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter slug: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.Delete(ctx, slug)
	return err
}

// GetArticleBySlug converts echo context to params.
func (w *ServerInterfaceWrapper) GetArticleBySlug(ctx *externalRef0.Ctx) error {
	var err error
	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameter("simple", false, "slug", ctx.Params("slug"), &slug)
	if err != nil {
		return externalRef0.NewError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter slug: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetArticleBySlug(ctx, slug)
	return err
}

// UpdateArticle converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateArticle(ctx *externalRef0.Ctx) error {
	var err error
	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameter("simple", false, "slug", ctx.Params("slug"), &slug)
	if err != nil {
		return externalRef0.NewError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter slug: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.UpdateArticle(ctx, slug)
	return err
}

// CreateAuthor converts echo context to params.
func (w *ServerInterfaceWrapper) CreateAuthor(ctx *externalRef0.Ctx) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.CreateAuthor(ctx)
	return err
}

// GetAuthorByName converts echo context to params.
func (w *ServerInterfaceWrapper) GetAuthorByName(ctx *externalRef0.Ctx) error {
	var err error
	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameter("simple", false, "name", ctx.Params("name"), &name)
	if err != nil {
		return externalRef0.NewError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter name: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetAuthorByName(ctx, name)
	return err
}

// GetTags converts echo context to params.
func (w *ServerInterfaceWrapper) GetTags(ctx *externalRef0.Ctx) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetTags(ctx)
	return err
}

type FiberRouter interface {
	Use(args ...interface{}) externalRef0.Router

	Get(path string, handlers ...externalRef0.Handler) externalRef0.Router
	Head(path string, handlers ...externalRef0.Handler) externalRef0.Router
	Post(path string, handlers ...externalRef0.Handler) externalRef0.Router
	Put(path string, handlers ...externalRef0.Handler) externalRef0.Router
	Delete(path string, handlers ...externalRef0.Handler) externalRef0.Router
	Connect(path string, handlers ...externalRef0.Handler) externalRef0.Router
	Options(path string, handlers ...externalRef0.Handler) externalRef0.Router
	Trace(path string, handlers ...externalRef0.Handler) externalRef0.Router
	Patch(path string, handlers ...externalRef0.Handler) externalRef0.Router

	Add(method, path string, handlers ...externalRef0.Handler) externalRef0.Router
	Static(prefix, root string, config ...externalRef0.Static) externalRef0.Router
	All(path string, handlers ...externalRef0.Handler) externalRef0.Router

	Group(prefix string, handlers ...externalRef0.Handler) externalRef0.Router

	Mount(prefix string, fiber *externalRef0.App) externalRef0.Router
}

// RegisterHandlers adds each server route to the FiberRouter.
func RegisterHandlers(router FiberRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router FiberRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.Add("GET", "/articles", func(ctx *externalRef0.Ctx) error {
		return wrapper.GetArticles(ctx)
	})

	router.Add("POST", "/articles", func(ctx *externalRef0.Ctx) error {
		return wrapper.CreateArticle(ctx)
	})

	router.Add("DELETE", "/articles/:slug", func(ctx *externalRef0.Ctx) error {
		return wrapper.Delete(ctx)
	})

	router.Add("GET", "/articles/:slug", func(ctx *externalRef0.Ctx) error {
		return wrapper.GetArticleBySlug(ctx)
	})

	router.Add("PUT", "/articles/:slug", func(ctx *externalRef0.Ctx) error {
		return wrapper.UpdateArticle(ctx)
	})

	router.Add("POST", "/authors", func(ctx *externalRef0.Ctx) error {
		return wrapper.CreateAuthor(ctx)
	})

	router.Add("GET", "/authors/:name", func(ctx *externalRef0.Ctx) error {
		return wrapper.GetAuthorByName(ctx)
	})

	router.Add("GET", "/tags", func(ctx *externalRef0.Ctx) error {
		return wrapper.GetTags(ctx)
	})

}