# restapi/types.go
// Package Swagger provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package restapi

import (
	"time"
)

// ArticleRequest defines model for ArticleRequest.
type ArticleRequest struct {
	AuthorId       *string   `json:"authorId,omitempty"`
	Body           *string   `json:"body,omitempty"`
	Description    *string   `json:"description,omitempty"`
	Favorited      *bool     `json:"favorited,omitempty"`
	FavoritesCount *int      `json:"favoritesCount,omitempty"`
	LikedUsers     *[]string `json:"likedUsers,omitempty"`
	Score          *float32  `json:"score,omitempty"`
	Slug           *string   `json:"slug,omitempty"`
	TagList        *[]string `json:"tagList,omitempty"`
	Title          *string   `json:"title,omitempty"`
}

// ArticleResponse defines model for ArticleResponse.
type ArticleResponse struct {
	Author         *AuthorResponse `json:"Author,omitempty"`
	AuthorID       *string         `json:"AuthorID,omitempty"`
	Body           *string         `json:"Body,omitempty"`
	CreatedAt      *time.Time      `json:"CreatedAt,omitempty"`
	Description    *string         `json:"Description,omitempty"`
	Favorited      *string         `json:"Favorited,omitempty"`
	FavoritesCount *int            `json:"FavoritesCount,omitempty"`
	ID             *string         `json:"ID,omitempty"`
	LikedUsers     *[]string       `json:"LikedUsers,omitempty"`
	Score          *float32        `json:"Score,omitempty"`
	Slug           *string         `json:"Slug,omitempty"`
	TagList        *[]TagResponse  `json:"TagList,omitempty"`
	Title          *string         `json:"Title,omitempty"`
	UpdatedAt      *time.Time      `json:"UpdatedAt,omitempty"`
}

// AuthorRequest defines model for AuthorRequest.
type AuthorRequest struct {
	Image      *string    `json:"image,omitempty"`
	Name       *string    `json:"name,omitempty"`
	RegisterOn *time.Time `json:"registerOn,omitempty"`
}

// AuthorResponse defines model for AuthorResponse.
type AuthorResponse struct {
	CreatedAt  *time.Time `json:"CreatedAt,omitempty"`
	ID         *string    `json:"ID,omitempty"`
	Image      *string    `json:"Image,omitempty"`
	Name       *string    `json:"Name,omitempty"`
	RegisterOn *time.Time `json:"RegisterOn,omitempty"`
	UpdatedAt  *time.Time `json:"UpdatedAt,omitempty"`
}

// TagResponse defines model for TagResponse.
type TagResponse struct {
	CreatedAt *time.Time `json:"CreatedAt,omitempty"`
	ID        *string    `json:"ID,omitempty"`
	Name      *string    `json:"Name,omitempty"`
	UpdatedAt *time.Time `json:"UpdatedAt,omitempty"`
}

// CreateArticleJSONBody defines parameters for CreateArticle.
type CreateArticleJSONBody ArticleRequest

// UpdateArticleJSONBody defines parameters for UpdateArticle.
type UpdateArticleJSONBody ArticleRequest

// CreateAuthorJSONBody defines parameters for CreateAuthor.
type CreateAuthorJSONBody AuthorRequest

// CreateArticleJSONRequestBody defines body for CreateArticle for application/json ContentType.
type CreateArticleJSONRequestBody CreateArticleJSONBody

// UpdateArticleJSONRequestBody defines body for UpdateArticle for application/json ContentType.
type UpdateArticleJSONRequestBody UpdateArticleJSONBody

// CreateAuthorJSONRequestBody defines body for CreateAuthor for application/json ContentType.
type CreateAuthorJSONRequestBody CreateAuthorJSONBody